Bring fixes from the dev branch into 6.2. We'll probably
need one final downmerge of dev into 6.2 before moving
over to a regular cherry-picking model.

+++ examples/multimediawidgets/player/playercontrols.h
@@ -68,13 +68,13 @@ public:
     explicit PlayerControls(QWidget *parent = nullptr);
 
     QMediaPlayer::PlaybackState state() const;
-    int volume() const;
+    float volume() const;
     bool isMuted() const;
     qreal playbackRate() const;
 
 public slots:
     void setState(QMediaPlayer::PlaybackState state);
-    void setVolume(int volume);
+    void setVolume(float volume);
     void setMuted(bool muted);
     void setPlaybackRate(float rate);
 
@@ -84,7 +84,7 @@ signals:
     void stop();
     void next();
     void previous();
-    void changeVolume(int volume);
+    void changeVolume(float volume);
     void changeMuting(bool muting);
     void changeRate(qreal rate);
